requirements.txt
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.7
djangorestframework==3.9.1
idna==2.8
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
pytz==2018.9
requests==2.21.0
uritemplate==3.0.0
urllib3==1.24.1
whitenoise==4.1.2
